Provide an independent trimmed spline.
At each end, the tube is bell shaped. This is apparent when any tube end is made very large.
This is a problem for modules that rely on radii along the spline. They expect a smooth radius variation.

Provide a spline that is trimmed at both ends by a specified number of points.