Add a subset of geometric shapes defined in parametric functions of VTK
 - Ellipsoid, Toroid, BohemianDome, Bour, Boy, ConicSpiral, CrossCap, Kuen,
 Mobius, PluckerConoid, Roman

They are all controlled by 4 markups points and UVW parameters. Some
shapes are further controlled by their own parameters.

The control points are used as follows:

 - p1: centre in Centered mode; opposite to p4 and through the centre in
 Circumferential mode
 - p2: X axis, radius, coefficient...
 - p3: Y axis, ...
 - p4: Z axis, ...; controls orientation also, like p1.

p2 and p3 are repositioned so that all points intersect at the centre at 90°.

The VTK parametric function is always computed at origin. A transform is applied
to the VTK actor to show the result within the control points.

The same approach is used in slice views, followed by mapping in each view.