Invert the colour schemes.
Label markups
Shape markups

 The GenericAnatomy colour table is used for the Label markups, and the Generic
table is used for the Shape markups. These are the same for the default labelmap
colour table node and for the default editor node. Inverting the scheme
for the markups may result in better distinction when segments use default
colours, at least for the Shape markups used as a tube around segments.