Account for the new 'SplineResolution' parameter.
Shape::Tube

The newly added SplineResolution parameter of the Tube shape impacts the default
values of the GetTrimmedSplineWorld() function. The latter function had a
default value of 25 for 'numberOfPointsToTrimAtStart' and for
'numberOfPointsToTrimAtEnd' while the resolution was hard coded at 100.

Since the minimum SplineResolution is 10, these default values no longer fit
the new situation. They are now set to -1, in which case, the number of points
corresponding to 25% of the spline resolution are trimmed for each parameter.
They can still be set to arbitrary positive values.