Add an option to choose the spline interval interpolation count.
Shape::Tube

From the original implementation, the spline points are interpolated by a factor
equal to the number of control point pairs of a tube. This means that the number
of point between each pair of control points is not equal to the spline
resolution.

In a previous commit, the spline resolution has become tunable and it is
reasonably expected to be equal to the number of spline points between each
control point pairs.

A new property, SplineNewInterpolationInterval, is added to choose the
number of interpolation intervals:

 - false: use the number of control point pairs, the default
 - true: use the number of control point pairs less one.